product price is updated with quantity selected

+++ lib/customerSide/view/pages/more_details_page.dart
@@ -110,7 +110,7 @@ class _MoreDetailsPageState extends State<MoreDetailsPage> {
               alignment: Alignment.bottomCenter,
               child: FilledButton(
                 onPressed: () {
-                  homePageController.cartAddBtnPressed(widget.itemIndex,buycount);
+                  homePageController.cartAddBtnPressed(widget.itemIndex,buycount.toDouble());
                   Get.to(CustomerMainPage());
                 },
                 child: Text(
